Change dominator_tree from <graph> to <graph_node>
Since we do not keep any references to the underlying graph beyond the
constructor, there is little point in requiring the type of the graph as
a template parameter. In fact, it may be difficult to determine the
exact type of a graph in this context, e.g., because a graph type may be
constructed with a lifetime shorter than that of the dominator_tree.